#19608 Move dependencies to the GIS bundle
